+++ src/errors/not-valid-object-id.error.ts
@@ -1,0 +1,8 @@
+import { ApiError } from './api.error';
+
+export class NotValidObjectIdError extends ApiError {
+
+    constructor() {
+        super('Not valid object id!', 400);
+    }
+}