+++ src/services/weather-stack.service.ts
@@ -1,6 +1,8 @@
 import { config } from '../config';
 import { logger } from '../logger';
 import { WsApiResponse } from '../types/ws-api.type';
+import { WsResponseCountryError } from '../errors/ws-response-country.error';
+import { ResponseError } from '../errors/response.error';
 
 export class WeatherStackService {
 
@@ -9,7 +11,7 @@ export class WeatherStackService {
             const response = await fetch(`http://api.weatherstack.com/current?access_key=${config?.wsApiAccessKey}&query=${city},${state},${zipCode}`);
 
             if (!response.ok) {
-                throw new Error(`HTTP error! Status: ${response.status}`);
+                throw new ResponseError(response.status);
             }
 
             logger.info(`Successfully received weather information for ${city}, ${state}, ${zipCode}.`);
@@ -17,7 +19,7 @@ export class WeatherStackService {
             const wsApiResponse = await response.json() as WsApiResponse;
 
             if (wsApiResponse.location.country !== 'United States of America') {
-                throw new Error(`Weather information was found for ${wsApiResponse.location.country} instead of US!`);
+                throw new WsResponseCountryError(wsApiResponse.location.country);
             }
 
             return wsApiResponse;